offtake: Sustainability Indices and Models for Wildlife Harvest and Hunting

Toolkit for assessing the sustainability of wildlife exploitation (bushmeat hunting, recreational hunting and fisheries). It implements the main index-based and model-based approaches catalogued by Adounke et al. (2026) doi:10.1016/j.biocon.2025.111606 and reviewed by Weinbaum et al. (2013) doi:10.1111/ele.12008 . Indices compare a harvest or abundance metric between hunted and reference sites: age structure comparison ('asc'), hunting yield comparison ('hyco') and population density comparison ('pdc'). Models compare observed offtake with an estimated maximum sustainable harvest: the production model of Robinson and Redford (1991) ('pro'), potential biological removal of Wade (1998) doi:10.1111/j.1748-7692.1998.tb00688.x ('pbr'), the maximum sustainable yield of the logistic model ('msy'), the spatially explicit source-sink model of Levi et al. (2009) ('biode') and a stochastic extension of PBR in the spirit of Manlik et al. (2022) doi:10.1111/cobi.13897 ('samse').

See also

Author

Maintainer: Stanislas Gandaho stangandaho@gmail.com (ORCID)